Feat: Add gene function discussion, experiment, and introduction prompts                                                                   │ │                                                                                                                                                               │ │   - Enhance gene function analysis with additional prompt sections                                                                                            │ │   - Add discussion, experiment, and introduction prompt handling                                                                                              │ │   - Update task manager timeout parameter type from int to float                                                                                              │ │   - Refactor gene function report generation with multi-step prompting                                                                                        │ │   - Add experiment protocol generation and integration

+++ src/mcp_server_phytomni/task_manager.py
@@ -169,7 +169,7 @@ async def update_task(url,
                       server_status: str,
                       server_file_path: str,
                       tool_result: str,
-                      timeout: int = 60,
+                      timeout: float = 60,
                       retriable_codes: List[int] = [429, 500, 502, 503, 504],
                       max_retries: int = 5):
     """Updates a task on a remote server.
